removed exisiting  folder as it is redundant
