--- time/timeTests/ProjectTimerTest.swift
@@ -1,9 +1,0 @@
-//
-//  ProjectTimerTest.swift
-//  dwyl time
-//
-//  Created by Sohil Pandya on 12/06/2017.
-//  Copyright © 2017 dwyl. All rights reserved.
-//
-
-import Foundation
